adds search bar draft and question componenets
